+++ Sources/Tessera/Internal/TesseraCanvasTile.swift
@@ -126,8 +126,9 @@ private extension TesseraCanvasTile {
     case var .organic(organicPlacement):
       organicPlacement.seed = seed
       return .organic(organicPlacement)
-    case .grid:
-      return configuration.placement
+    case var .grid(gridPlacement):
+      gridPlacement.seed = seed
+      return .grid(gridPlacement)
     }
   }
 
@@ -189,8 +190,8 @@ private extension TesseraCanvasTile {
     switch placement {
     case let .organic(organicPlacement):
       organicPlacement.seed
-    case .grid:
-      0
+    case let .grid(gridPlacement):
+      gridPlacement.seed
     }
   }